Added Url options for Content, Header, Footer, and Cover Added the missing Html options Made slightly more "sane" signature for Convert method Exposed a little more options for the Process updated Readme with samples/usage etc

+++ src/Shark.PdfConvert/PdfConvert.cs
@@ -40,9 +40,11 @@
             }
 
             // COVER
-            if (string.IsNullOrWhiteSpace(temporaryCoverFilePath) == false)
+            if (string.IsNullOrWhiteSpace(temporaryCoverFilePath) == false ||
+                string.IsNullOrWhiteSpace(config.PageCoverUrl) == false)
             {
-                options.AppendFormat(" cover  \"{0}\" ", temporaryCoverFilePath);
+                options.AppendFormat(" cover  \"{0}\" ",
+                    string.IsNullOrWhiteSpace(config.PageCoverUrl) ? temporaryCoverFilePath : config.PageCoverUrl);
 
                 if (string.IsNullOrWhiteSpace(config.CustomWkHtmlCoverArgs) == false)
                 {
@@ -53,9 +55,11 @@
             }
 
             // FOOTER
-            if (string.IsNullOrWhiteSpace(temporaryFooterFilePath) == false)
+            if (string.IsNullOrWhiteSpace(temporaryFooterFilePath) == false ||
+                string.IsNullOrWhiteSpace(config.PageFooterUrl) == false)
             {
-                options.AppendFormat(" --footer-html  \"{0}\" ", temporaryFooterFilePath);
+                options.AppendFormat(" --footer-html  \"{0}\" ",
+                    string.IsNullOrWhiteSpace(config.PageFooterUrl) ? temporaryFooterFilePath : config.PageFooterUrl);
 
                 if (string.IsNullOrWhiteSpace(config.CustomWkHtmlFooterArgs) == false)
                 {
@@ -66,9 +70,11 @@
             }
 
             // HEADER
-            if (string.IsNullOrWhiteSpace(temporaryHeaderFilePath) == false)
+            if (string.IsNullOrWhiteSpace(temporaryHeaderFilePath) == false ||
+                string.IsNullOrWhiteSpace(config.PageHeaderUrl) == false)
             {
-                options.AppendFormat(" --header-html  \"{0}\" ", temporaryHeaderFilePath);
+                options.AppendFormat(" --header-html  \"{0}\" ",
+                    string.IsNullOrWhiteSpace(config.PageHeaderUrl) ? temporaryHeaderFilePath : config.PageHeaderUrl);
 
                 if (string.IsNullOrWhiteSpace(config.CustomWkHtmlHeaderArgs) == false)
                 {
@@ -91,7 +97,9 @@
             }
 
             // PAGE
-            options.AppendFormat(" page \"{1}\" \"{0}\"", temporaryPdfFilePath, temporaryContentFilePath);
+            options.AppendFormat(" page \"{1}\" \"{0}\"",
+                temporaryPdfFilePath,
+                string.IsNullOrWhiteSpace(config.ContentUrl) ? temporaryContentFilePath : config.ContentUrl);
 
             // PAGE OPTIONS
             if (string.IsNullOrWhiteSpace(config.CustomWkHtmlPageArgs))
@@ -109,55 +117,40 @@
         }
 
         /// <summary>
-        /// Short cut method if you don't want to specify a stream for the content input, header, footer, etc
-        /// and use the PdfExportModel.Content member for your content instead of init'ing your own stream.
+        /// Converts an HTML source input into a PDF output using the WkHTMLToPDF tool.
+        /// The kitchen sink overload.
         /// </summary>
-        /// <param name="config"></param>
+        /// <param name="config">Conversion</param>
         /// <param name="pdfOutputStream"></param>
+        /// <param name="contentInputStream"></param>
+        /// <param name="coverInputStream"></param>
+        /// <param name="footerInputStream"></param>
+        /// <param name="headerInputStream"></param>
         /// <param name="outputCallback"></param>
         public static void Convert(PdfConversionSettings config,
-            Stream pdfOutputStream,
+            Stream pdfOutputStream = null,
+            Stream contentInputStream = null,
             Stream coverInputStream = null,
             Stream footerInputStream = null,
             Stream headerInputStream = null,
             Action<PdfConversionSettings, byte[]> outputCallback = null)
         {
-            using (var readerStream = new MemoryStream(Encoding.UTF8.GetBytes(config.Content)))
-            {
-                Convert(config, readerStream, pdfOutputStream, coverInputStream, footerInputStream, headerInputStream, outputCallback);
-            }
-        }
-
-        /// <summary>
-        /// Converts an HTML source input into a PDF output using the WkHTMLToPDF tool.
-        /// The kitchen sink overload.
-        /// </summary>
-        /// <param name="config">Conversion</param>
-        /// <param name="contentInputStream"></param>
-        /// <param name="pdfOutputStream"></param>
-        /// <param name="outputCallback"></param>
-        public static void Convert(PdfConversionSettings config,
-            Stream contentInputStream,
-            Stream pdfOutputStream,
-            Stream coverInputStream,
-            Stream footerInputStream,
-            Stream headerInputStream,
-            Action<PdfConversionSettings, byte[]> outputCallback)
-        {
             if (!File.Exists(config.PdfToolPath))
                 throw new ArgumentException($"File '{config.PdfToolPath}' not found. Check if wkhtmltopdf application is installed and set the correct location before calling this method.");
 
-            if (contentInputStream == null)
-                throw new ArgumentException("You must specify an input stream.");
+            if (contentInputStream == null &&
+                string.IsNullOrWhiteSpace(config.Content) &&
+                string.IsNullOrWhiteSpace(config.ContentUrl))
+                throw new ArgumentException("You must specify an input stream, static Content, or Content Url.");
 
-            if (pdfOutputStream == null)
-                throw new ArgumentException("You must specify an output stream.");
-
-            if (config.GenerateToc && coverInputStream == null)
-                throw new ArgumentException("You must specify a cover stream when specifying Generate TOC");
+            if (pdfOutputStream == null &&
+                string.IsNullOrWhiteSpace(config.OutputPath))
+                throw new ArgumentException("You must specify an output stream or an output path.");
 
             string temporaryContentFilePath = Path.Combine(config.TempFilesPath, $"{Guid.NewGuid()}.html");
-            string temporaryPdfFilePath = Path.Combine(config.TempFilesPath, $"{Guid.NewGuid()}.pdf");
+            string temporaryPdfFilePath = string.IsNullOrWhiteSpace(config.OutputPath) ?
+                Path.Combine(config.TempFilesPath, $"{Guid.NewGuid()}.pdf") :
+                config.OutputPath;
             string temporaryCoverFilePath = coverInputStream == null ?
                 null :
                 Path.Combine(config.TempFilesPath, $"{Guid.NewGuid()}.html");
@@ -182,9 +175,15 @@
                 {
                     using (var tempCoverFileStream = new FileStream(temporaryCoverFilePath, FileMode.CreateNew))
                     {
-                        coverInputStream.Position = 0;
+                        if (coverInputStream.CanSeek) coverInputStream.Position = 0;
                         coverInputStream.CopyTo(tempCoverFileStream);
                     }
+                }
+                else if (string.IsNullOrWhiteSpace(config.PageCoverUrl) &&
+                    string.IsNullOrWhiteSpace(config.PageCoverHtml) == false)
+                {
+                    // cover, if content is manually specified and no stream specified, then use the content
+                    File.WriteAllText(temporaryCoverFilePath, config.PageCoverHtml);
                 }
 
                 // FOOTER
@@ -192,11 +191,12 @@
                 {
                     using (var tempFooterFileStream = new FileStream(temporaryFooterFilePath, FileMode.CreateNew))
                     {
-                        footerInputStream.Position = 0;
+                        if (footerInputStream.CanSeek) footerInputStream.Position = 0;
                         footerInputStream.CopyTo(tempFooterFileStream);
                     }
                 }
-                else if (string.IsNullOrWhiteSpace(config.PageFooterHtml) == false)
+                else if (string.IsNullOrWhiteSpace(config.PageFooterUrl) &&
+                    string.IsNullOrWhiteSpace(config.PageFooterHtml) == false)
                 {
                     // footer, if content is manually specified and no stream specified, then use the content
                     File.WriteAllText(temporaryFooterFilePath, config.PageFooterHtml);
@@ -207,22 +207,36 @@
                 {
                     using (var tempHeaderFileStream = new FileStream(temporaryHeaderFilePath, FileMode.CreateNew))
                     {
-                        headerInputStream.Position = 0;
+                        if (headerInputStream.CanSeek) headerInputStream.Position = 0;
                         headerInputStream.CopyTo(tempHeaderFileStream);
                     }
                 }
-                else if (string.IsNullOrWhiteSpace(config.PageHeaderHtml) == false)
+                else if (string.IsNullOrWhiteSpace(config.PageHeaderUrl) &&
+                    string.IsNullOrWhiteSpace(config.PageHeaderHtml) == false)
                 {
                     // header, if content is manually specified and no stream specified, then use the content
                     File.WriteAllText(temporaryHeaderFilePath, config.PageHeaderHtml);
                 }
 
-
                 // CONTENT
-                using (var tempContentFileStream = new FileStream(temporaryContentFilePath, FileMode.CreateNew))
+                if (contentInputStream != null)
                 {
-                    contentInputStream.Position = 0;
-                    contentInputStream.CopyTo(tempContentFileStream);
+                    // use the stream specified
+                    using (var tempContentFileStream = new FileStream(temporaryContentFilePath, FileMode.CreateNew))
+                    {
+                        if (contentInputStream.CanSeek) contentInputStream.Position = 0;
+                        contentInputStream.CopyTo(tempContentFileStream);
+                    }
+                }
+                else if (string.IsNullOrWhiteSpace(config.ContentUrl) &&
+                  string.IsNullOrWhiteSpace(config.Content) == false)
+                {
+                    // use the content specified
+                    using (var tempContentFileStream = new FileStream(temporaryContentFilePath, FileMode.CreateNew))
+                    {
+                        byte[] contentAsBytes = Encoding.UTF8.GetBytes(config.Content);
+                        tempContentFileStream.Write(contentAsBytes, 0, contentAsBytes.Length);
+                    }
                 }
 
                 // start process
@@ -230,11 +244,11 @@
                 {
                     ProcessStartInfo startInfo = new ProcessStartInfo()
                     {
-                        CreateNoWindow = true,
+                        CreateNoWindow = config.ProcessOptionCreateNoWindow,
                         FileName = config.PdfToolPath,
                         Arguments = commandOptions,
-                        UseShellExecute = false,
-                        RedirectStandardError = true
+                        UseShellExecute = config.ProcessOptionUseShellExecute,
+                        RedirectStandardError = config.ProcessOptionRedirectStandardError
                     };
 
                     process.StartInfo = startInfo;
@@ -251,27 +265,29 @@
                             var error = startInfo.RedirectStandardError ?
                                 process.StandardError.ReadToEnd() :
                                 $"WkHTMLToPdf exited with code {process.ExitCode}.";
-                            throw new InvalidDataException($"WkHTMLToPdf conversion of '{config.OutputFilename}' failed. Output: \r\n{error}");
+                            throw new InvalidDataException($"WkHTMLToPdf conversion of HTML data failed. Output: \r\n{error}");
                         }
 
-                        throw new InvalidDataException($"WkHTMLToPdf conversion of '{config.OutputFilename}' failed. Output file '{temporaryPdfFilePath}' not found.");
+                        throw new InvalidDataException($"WkHTMLToPdf conversion of HTML data failed. Output file '{temporaryPdfFilePath}' not found.");
                     }
 
-                    using (Stream fs = new FileStream(temporaryPdfFilePath, FileMode.Open))
-                    {
-                        byte[] buffer = new byte[32 * 1024];
-                        int read;
+                    if (pdfOutputStream != null)
+                        using (Stream fs = new FileStream(temporaryPdfFilePath, FileMode.Open))
+                        {
+                            byte[] buffer = new byte[32 * 1024];
+                            int read;
 
-                        while ((read = fs.Read(buffer, 0, buffer.Length)) > 0)
-                            pdfOutputStream.Write(buffer, 0, read);
-                    }
+                            while ((read = fs.Read(buffer, 0, buffer.Length)) > 0)
+                                pdfOutputStream.Write(buffer, 0, read);
+                        }
 
                     outputCallback?.Invoke(config, File.ReadAllBytes(temporaryPdfFilePath));
                 }
             }
             finally
             {
-                if (File.Exists(temporaryPdfFilePath))
+                // if they specified an outputpath, don't kill the "temp" file
+                if (string.IsNullOrWhiteSpace(config.OutputPath) && File.Exists(temporaryPdfFilePath))
                     File.Delete(temporaryPdfFilePath);
                 if (File.Exists(temporaryContentFilePath))
                     File.Delete(temporaryContentFilePath);
